backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["tokitou-san <tokitousan2005@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
psycopg = "^3.1.13"
pillow = "^10.1.0"
django-dynamic-filenames = "^1.3.2"
django-cleanup = "^8.0.0"
python-dotenv = "^1.0.0"
black = "^23.11.0"
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
django-filter = "^23.4"
django-stubs-ext = "^4.2.5"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
